Presentation Powerpoint
Presentation Powerpoint is currently focused on our 450 presentation. The only difference will be that we will remove the methodologies slide and skip that part of the presentation. Additionally, we will focus on database related stuff while doing the demo use case to demonstrate how our database satisfies the requirements from the rubric.